requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
qrcode[pil]==7.4.2
python-decouple==3.8
psycopg2-binary==2.9.9
django-extensions==3.2.3
celery==5.3.4
redis==5.0.1
reportlab==4.0.7
pandas==2.1.4
numpy==1.25.2
django-filter==23.5
gunicorn==21.2.0
whitenoise==6.6.0
